requirements.txt
streamlit>=1.28.0
google-cloud-aiplatform>=1.38.0
google-cloud-aiplatform[langchain]>=1.35.0
pandas>=1.5.0
numpy>=1.24.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
streamlit-authenticator>=0.2.3
plotly>=5.15.0
streamlit-option-menu>=0.3.6
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
langchain>=0.1.0
langchain-google-vertexai>=1.0.0
vertexai>=1.71.1
google>= 3.0.0
dotenv>=0.9.9
vertexai>=1.38.0
googlemaps>=4.10.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
fpdf2>=2.7.6
ics>=0.7.2
requests>=2.31.0
typing-extensions>=4.8.0
